remove duplicate pushback code
