+++ fbw-a380x/src/wasm/systems/a380_systems/src/hydraulic/mod.rs
@@ -3,7 +3,7 @@ use nalgebra::Vector3;
 use std::time::Duration;
 use uom::si::{
     acceleration::meter_per_second_squared,
-    angle::{degree, radian},
+    angle::degree,
     angular_velocity::{radian_per_second, revolution_per_minute},
     electric_current::ampere,
     f64::*,
@@ -39,6 +39,7 @@ use systems::{
             SteeringRatioToAngle,
         },
         pumps::PumpCharacteristics,
+        pushback::PushbackTug,
         trimmable_horizontal_stabilizer::{
             TrimmableHorizontalStabilizerActuator, TrimmableHorizontalStabilizerMotorController,
         },
@@ -49,12 +50,11 @@ use systems::{
     landing_gear::{GearSystemSensors, LandingGearControlInterfaceUnitSet, TiltingGear},
     overhead::{AutoOffFaultPushButton, AutoOnFaultPushButton},
     shared::{
-        interpolation, low_pass_filter::LowPassFilter, random_from_range,
-        update_iterator::MaxStepLoop, AdirsDiscreteOutputs, AirbusElectricPumpId,
-        AirbusEngineDrivenPumpId, DelayedFalseLogicGate, DelayedPulseTrueLogicGate,
-        DelayedTrueLogicGate, ElectricalBusType, ElectricalBuses, EngineFirePushButtons, GearWheel,
-        HydraulicColor, LandingGearHandle, LgciuInterface, LgciuWeightOnWheels,
-        ReservoirAirPressure, SectionPressure,
+        interpolation, random_from_range, update_iterator::MaxStepLoop, AdirsDiscreteOutputs,
+        AirbusElectricPumpId, AirbusEngineDrivenPumpId, DelayedFalseLogicGate,
+        DelayedPulseTrueLogicGate, DelayedTrueLogicGate, ElectricalBusType, ElectricalBuses,
+        EngineFirePushButtons, GearWheel, HydraulicColor, LandingGearHandle, LgciuInterface,
+        LgciuWeightOnWheels, ReservoirAirPressure, SectionPressure,
     },
     simulation::{
         InitContext, Read, Reader, SimulationElement, SimulationElementVisitor, SimulatorReader,
@@ -4041,87 +4041,6 @@ impl SimulationElement for A380BrakingForce {
     }
 }
 
-struct PushbackTug {
-    nw_strg_disc_memo_id: VariableIdentifier,
-    state_id: VariableIdentifier,
-    steer_angle_id: VariableIdentifier,
-
-    steering_angle_raw: Angle,
-    steering_angle: LowPassFilter<Angle>,
-
-    // Type of pushback:
-    // 0 = Straight
-    // 1 = Left
-    // 2 = Right
-    // 3 = Assumed to be no pushback
-    // 4 = might be finishing pushback, to confirm
-    state: f64,
-    nose_wheel_steering_pin_inserted: DelayedFalseLogicGate,
-}
-impl PushbackTug {
-    const DURATION_AFTER_WHICH_NWS_PIN_IS_REMOVED_AFTER_PUSHBACK: Duration =
-        Duration::from_secs(15);
-
-    const STATE_NO_PUSHBACK: f64 = 3.;
-
-    const STEERING_ANGLE_FILTER_TIME_CONSTANT: Duration = Duration::from_millis(1500);
-
-    fn new(context: &mut InitContext) -> Self {
-        Self {
-            nw_strg_disc_memo_id: context.get_identifier("HYD_NW_STRG_DISC_ECAM_MEMO".to_owned()),
-            state_id: context.get_identifier("PUSHBACK STATE".to_owned()),
-            steer_angle_id: context.get_identifier("PUSHBACK ANGLE".to_owned()),
-
-            steering_angle_raw: Angle::default(),
-            steering_angle: LowPassFilter::new(Self::STEERING_ANGLE_FILTER_TIME_CONSTANT),
-
-            state: Self::STATE_NO_PUSHBACK,
-            nose_wheel_steering_pin_inserted: DelayedFalseLogicGate::new(
-                Self::DURATION_AFTER_WHICH_NWS_PIN_IS_REMOVED_AFTER_PUSHBACK,
-            ),
-        }
-    }
-
-    fn update(&mut self, context: &UpdateContext) {
-        self.nose_wheel_steering_pin_inserted
-            .update(context, self.is_pushing());
-
-        if self.is_pushing() {
-            self.steering_angle
-                .update(context.delta(), self.steering_angle_raw);
-        } else {
-            self.steering_angle.reset(Angle::default());
-        }
-    }
-
-    fn is_pushing(&self) -> bool {
-        (self.state - PushbackTug::STATE_NO_PUSHBACK).abs() > f64::EPSILON
-    }
-}
-impl Pushback for PushbackTug {
-    fn is_nose_wheel_steering_pin_inserted(&self) -> bool {
-        self.nose_wheel_steering_pin_inserted.output()
-    }
-
-    fn steering_angle(&self) -> Angle {
-        self.steering_angle.output()
-    }
-}
-impl SimulationElement for PushbackTug {
-    fn read(&mut self, reader: &mut SimulatorReader) {
-        self.state = reader.read(&self.state_id);
-
-        self.steering_angle_raw = Angle::new::<radian>(reader.read(&self.steer_angle_id));
-    }
-
-    fn write(&self, writer: &mut SimulatorWriter) {
-        writer.write(
-            &self.nw_strg_disc_memo_id,
-            self.is_nose_wheel_steering_pin_inserted(),
-        );
-    }
-}
-
 /// Autobrake controller computes the state machine of the autobrake logic, and the deceleration target
 /// that we expect for the plane
 pub struct A380AutobrakeController {
@@ -6507,7 +6426,7 @@ mod tests {
         };
 
         use uom::si::{
-            angle::degree,
+            angle::{degree, radian},
             electric_potential::volt,
             length::foot,
             ratio::{percent, ratio},
