+++ fbw-common/src/wasm/systems/systems/src/hydraulic/mod.rs
@@ -42,6 +42,7 @@ pub mod landing_gear;
 pub mod linear_actuator;
 pub mod nose_steering;
 pub mod pumps;
+pub mod pushback;
 pub mod rudder_control;
 pub mod trimmable_horizontal_stabilizer;
